Cleaned and performed troubleshooting.
Various points of logic in the game (especially lonewolf to house_bporch) conflicted and were accidentally running all at once. I attempted to resolve this as best I could.

The house, I'm afraid, may be problematic for players who get creative. If there is a location change, the option-based condition may be met and the next part of the script will fire prematurely. This should be considered in moving forward.