Version 0.2.0 (#4)
* Fixed some minor progression issues.

I also attempted to put  in place for a rudimentary Game Over screen that allows you to replay the game, but it isn't working quite as expected so far.

* Updated README.md

* Fixed grammar, code format/spacing, and did troubleshooting.

I cleaned up some of the dialogue and the way the engine calls new parts of the dialogue.

I have decided that the code is most consistent when I do three returns after each section, but just two returns for different options inside the section.

I found out the entire living room was pulling from house_kitchen, and the wrong variable caused the game to prematurely end.

* Solved the Game Over issue.

The gameover issue was solved from two different standpoints. I captured the y/n inputs themselves and added an if/then to prevent it from execution the main game option function.

After doing so, I then discovered that the issue was actually caused by the `<p></p>` options I'd left as placeholders. Javascript would attempt to run my dirty refresh, but instead was pulled to another script altogether.

Co-authored-by: TheCzar <Admin@czarsaparilla.sturmhaus.net>