Javascript is almost the dominant language.
jQuery is still controlling the input, but the game is flowing again with the innerHTML set up as well. I hope that in the next push, I'll completely remove jQuery and its library from the game, as it's not been necessary, but I was using a YouTube tutorial to build that part of it.